replace text with icons in nav bar
