+++ app/Root.js
@@ -1,49 +1,68 @@
+'use strict';
 import React, { Component } from 'react';
-import styles from './styles/main';
-import ContactCard from './components/ContactCard';
-import { View, Text, Image, TouchableHighlight } from 'react-native';
-import Main from './screens/Main';
-import { Actions } from 'react-native-router-flux';
-// import firebase, { contactsRef, provider } from '../firebase';
+import ContactList from './screens/ContactList';
+import AddContact from './screens/AddContact';
+import IndividualContact from './screens/IndividualContact';
+import { Text } from 'react-native';
+import { Router, Scene, ActionConst } from 'react-native-router-flux';
+import Icon from 'react-native-vector-icons/Ionicons';
 
-export default class birthdayTracker extends Component {
-  // constructor() {
-  //   super();
-  //   this.state = {
-  //     user: null,
-  //   };
-  // }
-  //
-  // componentDidMount() {
-  //   firebase.auth().onAuthStateChanged((user) => {
-  //     this.setState({ user });
-  //   });
-  // }
-  //
-  // triggerSignOut() {
-  //   firebase.auth().signOut();
-  //   this.setState({ user: null });
-  // }
-
-  render() {
-    // const { user } = this.state;
-    // if(user) {
-    //   return (
-    //     <Main />
-    //   );
-    // }
-    return (
-      <Main />
-      // <View style={styles.loginScreen}>
-      //   <Text style={styles.appTitle}>Saves the Day</Text>
-      //   <Image source={require('./img/present.png')} />
-      //   <TouchableHighlight
-      //     style={styles.loginButton}
-          // onPress={() => firebase.auth().signInWithCredential(provider.credential(R63gFHdzcXy14dc1tQhnnbxi))}
-      //     underlayColor='#7ac4ba'>
-      //     <Text style={styles.loginButtonText}>Login with Google</Text>
-      //   </TouchableHighlight>
-      // </View>
-    );
-  }
+const TabIcon = ({ selected, title }) => {
+  return (
+    <Text>{title}</Text>
+  );
 }
+
+const ContactBook = ({ selected }) => {
+  return (
+    <Icon name='ios-book' size={30} color='#222222' />
+  );
+}
+
+const PlusContact = ({ selected }) => {
+  return (
+    <Icon name='md-person-add' size={30} color='#222222' />
+  );
+}
+
+const Root = () => {
+
+  return (
+    <Router>
+      <Scene key="root">
+
+        <Scene
+          key="tabbar"
+          tabs={true}
+          tabBarStyle={{ backgroundColor: '#EAE8E8' }}
+        >
+
+          <Scene key="contactList" title="Contact List" icon={ContactBook}>
+            <Scene
+              key="contacts"
+              component={ContactList}
+              title="Saves the Day"
+            />
+          </Scene>
+
+          <Scene key="addContacts" title="Add Contact" icon={PlusContact}>
+            <Scene
+              key="individualContact"
+              component={AddContact}
+              title="Add Contact"
+            />
+          </Scene>
+
+          <Scene
+            key="contactInfo"
+            component={IndividualContact}
+            title="Contact Info"
+          />
+        </Scene>
+
+      </Scene>
+    </Router>
+  );
+}
+
+export default Root;
