--- app/screens/Main.js
@@ -1,54 +1,0 @@
-'use strict';
-import React, { Component } from 'react';
-import ContactList from './ContactList';
-import AddContact from './AddContact';
-import IndividualContact from './IndividualContact';
-import { Text } from 'react-native';
-import { Router, Scene, ActionConst } from 'react-native-router-flux';
-
-const TabIcon = ({ selected, title }) => {
-  return (
-    <Text>{title}</Text>
-  );
-}
-
-const Main = () => {
-  return (
-    <Router>
-      <Scene key="root">
-
-        <Scene
-          key="tabbar"
-          tabs={true}
-          tabBarStyle={{ backgroundColor: '#EAE8E8' }}
-        >
-
-          <Scene key="contactList" title="Contact List" icon={TabIcon}>
-            <Scene
-              key="contacts"
-              component={ContactList}
-              title="Saves the Day"
-            />
-          </Scene>
-
-          <Scene key="addContacts" title="Add Contact" icon={TabIcon}>
-            <Scene
-              key="individualContact"
-              component={AddContact}
-              title="Add Contact"
-            />
-          </Scene>
-
-          <Scene
-            key="contactInfo"
-            component={IndividualContact}
-            title="Contact Info"
-          />
-        </Scene>
-
-      </Scene>
-    </Router>
-  );
-}
-
-export default Main;
