+++ app/styles/Button.scss
@@ -1,8 +1,8 @@
 button {
   background-color: #F7E74A;
   border-radius: 5px;
-  margin: 20;
-  padding: 5;
+  margin: 10px;
+  padding: 5px;
   width: 80px;
   shadow-color: #000000;
   shadow-opacity: 0.3;
