+++ app/styles/main.js
@@ -1,1 +1,1 @@
-module.exports = require('react-native').StyleSheet.create({"addContact":{"flex":1,"backgroundColor":"#09C6AB","alignItems":"center","paddingTop":75},"firstNameInput":{"backgroundColor":"#EAE8E8","borderRadius":10,"marginBottom":10,"width":320},"lastNameInput":{"backgroundColor":"#EAE8E8","borderRadius":10,"marginBottom":10,"width":320},"inputFields":{"color":"#02556D","height":40,"textAlign":"center"},"datePicker":{"borderRadius":10,"backgroundColor":"#EAE8E8","width":320},"avatar":{"borderRadius":62,"height":125,"marginBottom":10,"width":125},"loginScreen":{"alignItems":"center","backgroundColor":"#068888","flex":1,"paddingTop":90},"loginButton":{"backgroundColor":"#09C6AB","marginTop":30,"padding":10,"width":250},"loginButtonText":{"color":"#02556D","fontSize":16,"textAlign":"center"},"appTitle":{"color":"#F7E74A","fontSize":40,"fontWeight":"bold","marginBottom":30,"textAlign":"center"},"contactList":{"alignItems":"center","backgroundColor":"#09C6AB","flex":1,"justifyContent":"center","paddingBottom":60},"searchBar":{"backgroundColor":"#EAE8E8","marginTop":80,"marginRight":0,"marginBottom":10,"marginLeft":0,"width":320,"borderRadius":10},"searchInput":{"color":"#02556D","height":40,"textAlign":"center"},"button":{"backgroundColor":"#F7E74A","borderRadius":5,"margin":20,"padding":5,"width":80,"shadowColor":"#000000","shadowOpacity":0.3,"shadowRadius":3},"button shadow-offset":{"height":0,"width":0},"buttonText":{"color":"#02556D","fontSize":16,"textAlign":"center"},"contactCard":{"backgroundColor":"#09C6AB","borderBottomWidth":0.5,"borderColor":"#026d8b","flex":1,"paddingTop":10,"paddingRight":0,"paddingBottom":10,"paddingLeft":10,"width":320},"nameText":{"color":"#013240","fontSize":16,"fontWeight":"bold","paddingTop":0,"paddingRight":0,"paddingBottom":5,"paddingLeft":20},"birthdayText":{"color":"#013240"},"row":{"flexDirection":"row"},"card":{"flex":1,"justifyContent":"center","marginLeft":20},"listAvatar":{"borderRadius":37,"height":75,"width":75},"individualContact":{"flex":1,"backgroundColor":"#09C6AB","alignItems":"center"},"notesField":{"backgroundColor":"#EAE8E8","borderRadius":5,"marginTop":20,"width":200},"notesInput":{"color":"#02556D","paddingTop":2,"paddingRight":10,"paddingBottom":10,"paddingLeft":10},"individualAvatar":{"borderRadius":62,"height":125,"marginTop":100,"marginRight":0,"marginBottom":10,"marginLeft":0,"width":125}});
+module.exports = require('react-native').StyleSheet.create({"addContact":{"flex":1,"backgroundColor":"#09C6AB","alignItems":"center","paddingTop":75},"firstNameInput":{"backgroundColor":"#EAE8E8","borderRadius":10,"marginBottom":10,"width":320},"lastNameInput":{"backgroundColor":"#EAE8E8","borderRadius":10,"marginBottom":10,"width":320},"inputFields":{"color":"#02556D","height":40,"textAlign":"center"},"datePicker":{"borderRadius":10,"backgroundColor":"#EAE8E8","width":320},"avatar":{"borderRadius":62,"height":125,"marginBottom":10,"width":125},"loginScreen":{"alignItems":"center","backgroundColor":"#068888","flex":1,"paddingTop":90},"loginButton":{"backgroundColor":"#09C6AB","marginTop":30,"padding":10,"width":250},"loginButtonText":{"color":"#02556D","fontSize":16,"textAlign":"center"},"appTitle":{"color":"#F7E74A","fontSize":40,"fontWeight":"bold","marginBottom":30,"textAlign":"center"},"contactList":{"alignItems":"center","backgroundColor":"#09C6AB","flex":1,"justifyContent":"center","paddingBottom":60},"searchBar":{"backgroundColor":"#EAE8E8","marginTop":80,"marginRight":0,"marginBottom":10,"marginLeft":0,"width":320,"borderRadius":10},"searchInput":{"color":"#02556D","height":40,"textAlign":"center"},"button":{"backgroundColor":"#F7E74A","borderRadius":5,"margin":10,"padding":5,"width":80,"shadowColor":"#000000","shadowOpacity":0.3,"shadowRadius":3},"button shadow-offset":{"height":0,"width":0},"buttonText":{"color":"#02556D","fontSize":16,"textAlign":"center"},"contactCard":{"backgroundColor":"#09C6AB","borderBottomWidth":0.5,"borderColor":"#026d8b","flex":1,"paddingTop":10,"paddingRight":0,"paddingBottom":10,"paddingLeft":10,"width":320},"nameText":{"color":"#013240","fontSize":16,"fontWeight":"bold","paddingTop":0,"paddingRight":0,"paddingBottom":5,"paddingLeft":20},"birthdayText":{"color":"#013240"},"row":{"flexDirection":"row"},"card":{"flex":1,"justifyContent":"center","marginLeft":20},"listAvatar":{"borderRadius":37,"height":75,"width":75},"individualContact":{"flex":1,"backgroundColor":"#09C6AB","alignItems":"center"},"notesField":{"backgroundColor":"#EAE8E8","borderRadius":5,"marginTop":20,"width":200},"notesInput":{"color":"#02556D","paddingTop":2,"paddingRight":10,"paddingBottom":10,"paddingLeft":10},"individualAvatar":{"borderRadius":62,"height":125,"marginTop":100,"marginRight":0,"marginBottom":10,"marginLeft":0,"width":125}});
