+++ package.json
@@ -10,6 +10,7 @@
     "firebase": "^3.1.0",
     "node-sass": "^3.10.1",
     "react": "15.3.2",
+    "react-dom": "^15.3.2",
     "react-native": "0.35.0",
     "react-native-autogrow-textinput": "^2.1.1",
     "react-native-datepicker": "^1.3.2",
@@ -32,7 +33,6 @@
     "mocha": "^3.1.2",
     "mockdate": "^2.0.0",
     "react-addons-test-utils": "^15.3.2",
-    "react-dom": "^15.3.2",
     "react-test-renderer": "15.3.2"
   }
 }
